test: split selection tests out of the int8 export suite (Codacy file NLOC)
The merge-readiness tests pushed test_export_grok1_int8_npy.py to 574 NLOC,
over Codacy's 500-per-file limit — the one issue its PR check reported while the
CLI was clean per-file. The selection/manifest-validation classes exercise
export_grok1_int8_select, already a separate module from the pickle scanner, so
that was the natural seam: they move to test_export_grok1_int8_select.py and are
wired into the Python-scripts workflow.

No test content changed; both files are now well under the limit and lizard is
clean at every Codacy threshold (function CCN 8, function NLOC 50, params 8,
file NLOC 500). 116 tests still pass.
